Update Create Dataset Study tab
Functionality for the Create Dataset Study tab has been added
according to slide 4 of the submission wizard mockup slides.

Two model classes: Dataset and Images have new functions added.

DatasetSubmissionController class has new getDataset() and
isSubmitter() added. The actionCreate1() function has been
updated to provide image functionality when using
no_image.jpg or is set is_no_image.

The datasetSubmission view class, create1 has a link to CKEditor
version 4.14 added to it so it is loaded from a CDN. Javascript
initialisation code for CKEditor added so it is used to write a
rich text description for the dataset.

A images/uploads directory has been created with a .gitkeep file in
it to preserve its presence in github. This directory is used for
the image uploading functionality.

Resolves WL Sprint 1 task:
https://jirasoft.ffflabel-dev.com/browse/GIG-4 for issue #241.