Fix Study tab re-direct problem
When other dataset submission wizard tabs are clicked before the
Study tab information has been saved, the wizard goes to the
"/user/view_profile" page. This is a bug and the re-direct should
go back to the Study tab at "/datasetSubmission/create1". This fix
has been implemented in this commit.