Update Study tab using "datasetsubmission/datasetmanagement" page
A new dataset is created by using the
"datasetsubmission/create1" page. After the Study
tab is saved then it re-directs to
"datasetsubmission/datasetmanagement" page.

From looking at the WL code, this appears to have
been implemented as follows. The actionCreate1()
function has been updated to check if the dataset
being created is a test submission or not. If the
dataset has an internal ID then it calls
actionDatasetManagement() which is a new function
that fetches the dataset model for the ID, manages
the image for the dataset and renders the Study
tab. If a new dataset is being created then it
renders an empty Study tab.

The actionSampleManagement() function has been
updated to work with a test dataset submission.

The _tabs_navigation.php script now calls
actionDatasetManagement() in
DatasetSubmissionController to update an existing
dataset.

DatasetSubmissionController class seems more
complicated with the changes made here from what
I understand in the WL code.

This commit is a port of f015b5af "Return
datasetManagement page for updating Study Tab"
from lump1k-sergey branch.