Check dataset if it is a test submission and re-direct
New getRedirectUrl() function has been added to
DatasetSubmissionController which returns
"/datasetSubmission/create1". This function is now called anywhere
in DatasetSubmissionController which used to redirect to
"/datasetSubmission/create1".

In _tabs_navigation.php, the dataset is checked to see if it a test
submission and then re-directed to its test dataset pages.

This commit is a port of ad18104b Bug fix. Test dataset when click
to another tab redirect to test datase(t) from lump1k-sergey branch.