Change select and confirm messages in Sample tab
The drop down menu for choosing a template informs the user to select a
template. There is also a confirmatory message displayed saying that
data will be overwritten when using a new template.

This commit is a port of 240bc30a Bug fix. When set template change
confirm message from lump1k-sergey branch.