Display scroll bar in Sample tab table
In case there are extra columns then a scroll bar can be used to view
all the columns. Implemented by adding an overflow property with auto
as the value to the sample table.

This is a port of cf4a29eb Bug Fix. Title for delete column wasnt
visible commit from lump1k-sergey branch.