Add console command for creating attributes
An executable console command has been added to create new attributes
if they are not already present in GigaDB. It uses a new function
findByAttrName($attributeName) in Attribute class that is used to
check if these attributes are already present or not. Other classes
now use findByAttrName($attributeName).

actionValidateSamples() has been updated in DatasetSubmissionController
so that it uses lower case to match attribute_name values.

This commit is a port of cd551af0 Create command for addting test
attributes for test template from lump1k-sergey branch.