Fix extra column bug in table in Sample tab
sampleManagement class has been updated so that when a new sample row is
added, only 3 columns for Sample ID, Species name and Description are
created in the table in Sample tab. Previously, an extra 4th column was
being created which is a bug.

Function actionSaveSamples() in DatasetSubmissionController has been
updated to use findByAttrName() to look up attributes. Sample deletion
in this function is now wrapped in a try-catch block too.

This commit is a port of 905227b0 Remove extra column in Sample Tab from
lump1k-sergey branch.