Fix error on Sample tab
The Autocomplete for species functionality has been removed because the
species table contains more than 1 million rows and timeout error
happens. This was achieved by updating the function
actionSampleManagement() so that all species (and attributes) are not
returned and passed to be rendered by sampleManagement view anymore.The
attributes autocomplete functionality has been reimplemented as an ajax
call in sampleManagement.php.

This commit is a port of c4a16ac2 Fix error on Sample tab from
lump1k-sergey branch.