Fix units bug for attributes in Sample tab
The JS function getAttributesAutocomplete() in sampleManagement.php has
been updated to call checkUnit() which is a new JS function that has
also been added to sampleManagement.php. checkUnit() wraps the ajax code
calling the actionCheckUnit() in DatasetSubmissionController via
/datasetSubmission/checkUnit URL as a JS function.

Function actionSaveSamples() has been updated in
DatasetSubmissionController. Not sure what is going on as the code has
no comments but looks like how it deals with attributes has changed.

This commit is a port of 49cbf864 Fix bug with units for attribute in
Sample tab from lump1k-sergey branch.