Add checkbox to use generic image for dataset
If no image file is selected when creating new dataset then the
Image model class that has been updated for the WL submission
wizard will complain about this. A checkbox to use a generic image
has been added to the create dataset page which will use this
image file for the dataset.