Fix(StoredDatasetConnections): Add clause to catch 504 time out errors
Retrieving citation information sometimes results in 504 gateway time out errors
and dataset web page is not displayed. A catch clause has been added to detect
this error with a null value given for the citation which allows the dataset page to
be shown.

Changes to the component classes:

StoredDatasetConnections.php has been updated to include a clause to catch
all 500-level errors which includes the 504 Gateway timeout error sometimes
returned by dataset pages such as parrot when citation information cannot be
retrieved.

Changes to the tests:

gigadb_testdata.pgdmp file has been updated to include the Puerto Rican parrot
data. Any dataset requiring retrieval of citation data could have been used by
testItShouldDisplayCitations() but it is usually this parrot dataset that results in a
504 error in failed acceptance tests.

DatasetViewTest.php contains a new functional test testItShouldDisplayCitations()
to test that citation information is present for the parrot dataset page at
http://gigadb.dev/dataset/100039.

RSSFeedTest.php has been updated to include the identifier for the parrot dataset
page.

The testItShouldShowLatestNews() function in protected/tests/functional/SiteTest.php
has been updated to include the date (2012-09-11) for the parrot dataset page.

A new unit test testStoredReturnsPublicationsWithGatewayTimeoutError() has been
added to protected/tests/unit/StoredDatasetConnectionsTest. This test simulates
two 504 gateway timeout errors which are caught by the ServerException catch
block when getPublications() in StoredDatasetConnections is called. The value for
the citations key value pair is null when there is a 504 error returned by https://doi.org
and it is this null citations value that the testStoredReturnsPublicationsWithGatewayTimeoutError()
checks for.

Closes #537